nio encoder/decoder
