+++ Sources/PostgreSQL/Database/PostgreSQLDatabaseConfig.swift
@@ -9,7 +9,7 @@ public struct PostgreSQLDatabaseConfig {
     public let hostname: String
 
     /// Destination port.
-    public let port: UInt16
+    public let port: Int
 
     /// Username to authenticate.
     public let username: String
@@ -22,7 +22,7 @@ public struct PostgreSQLDatabaseConfig {
     public let password: String?
     
     /// Creates a new `PostgreSQLDatabaseConfig`.
-    public init(hostname: String, port: UInt16, username: String, database: String? = nil, password: String? = nil) {
+    public init(hostname: String, port: Int, username: String, database: String? = nil, password: String? = nil) {
         self.hostname = hostname
         self.port = port
         self.username = username
