--- Sources/PostgreSQL/PostgreSQLMessageDecoder.swift
@@ -1,8 +1,0 @@
-//
-//  PostgreSQLMessageDecoder.swift
-//  PostgreSQL
-//
-//  Created by Tanner on 3/2/18.
-//
-
-import Foundation
