+++ Tests/PostgreSQLTests/PostgreSQLConnectionTests.swift
@@ -2,18 +2,17 @@ import Async
 import Foundation
 import XCTest
 import PostgreSQL
-import TCP
 
 class PostgreSQLConnectionTests: XCTestCase {
     func testVersion() throws {
-        let (client, eventLoop) = try PostgreSQLConnection.makeTest()
-        let results = try client.simpleQuery("SELECT version();").await(on: eventLoop)
+        let client = try PostgreSQLConnection.makeTest()
+        let results = try client.simpleQuery("SELECT version();").wait()
         try XCTAssert(results[0]["version"]?.decode(String.self).contains("10.") == true)
     }
 
     func testSelectTypes() throws {
-        let (client, eventLoop) = try PostgreSQLConnection.makeTest()
-        let results = try client.query("select * from pg_type;").await(on: eventLoop)
+        let client = try PostgreSQLConnection.makeTest()
+        let results = try client.query("select * from pg_type;").wait()
         if results.count > 350 {
             let name = try results[128]["typname"]?.decode(String.self)
             XCTAssert(name != "")
@@ -23,11 +22,11 @@ class PostgreSQLConnectionTests: XCTestCase {
     }
 
     func testParse() throws {
-        let (client, eventLoop) = try PostgreSQLConnection.makeTest()
+        let client = try PostgreSQLConnection.makeTest()
         let query = """
         select * from "pg_type" where "typlen" = $1 or "typlen" = $2
         """
-        let rows = try client.query(query, [1, 2]).await(on: eventLoop)
+        let rows = try client.query(query, [1, 2]).wait()
 
         for row in rows {
             try XCTAssert(row["typlen"]?.decode(Int.self) == 1 || row["typlen"]?.decode(Int.self) == 2)
@@ -35,7 +34,7 @@ class PostgreSQLConnectionTests: XCTestCase {
     }
 
     func testTypes() throws {
-        let (client, eventLoop) = try PostgreSQLConnection.makeTest()
+        let client = try PostgreSQLConnection.makeTest()
         let createQuery = """
         create table kitchen_sink (
             "smallint" smallint,
@@ -67,8 +66,8 @@ class PostgreSQLConnectionTests: XCTestCase {
             -- "uuid" uuid
         );
         """
-        _ = try client.query("drop table if exists kitchen_sink;").await(on: eventLoop)
-        let createResult = try client.query(createQuery).await(on: eventLoop)
+        _ = try client.query("drop table if exists kitchen_sink;").wait()
+        let createResult = try client.query(createQuery).wait()
         XCTAssertEqual(createResult.count, 0)
 
         let insertQuery = """
@@ -102,9 +101,9 @@ class PostgreSQLConnectionTests: XCTestCase {
             -- "uuid" uuid
         );
         """
-        let insertResult = try! client.query(insertQuery).await(on: eventLoop)
+        let insertResult = try client.query(insertQuery).wait()
         XCTAssertEqual(insertResult.count, 0)
-        let queryResult = try client.query("select * from kitchen_sink").await(on: eventLoop)
+        let queryResult = try client.query("select * from kitchen_sink").wait()
         if queryResult.count == 1 {
             let row = queryResult[0]
             try XCTAssertEqual(row["smallint"]?.decode(Int16.self), 1)
@@ -128,7 +127,7 @@ class PostgreSQLConnectionTests: XCTestCase {
     }
 
     func testParameterizedTypes() throws {
-        let (client, eventLoop) = try PostgreSQLConnection.makeTest()
+        let client = try PostgreSQLConnection.makeTest()
         let createQuery = """
         create table kitchen_sink (
             "smallint" smallint,
@@ -161,8 +160,8 @@ class PostgreSQLConnectionTests: XCTestCase {
             -- "bit" bit(16),
         );
         """
-        _ = try client.query("drop table if exists kitchen_sink;").await(on: eventLoop)
-        let createResult = try client.query(createQuery).await(on: eventLoop)
+        _ = try client.query("drop table if exists kitchen_sink;").wait()
+        let createResult = try client.query(createQuery).wait()
         XCTAssertEqual(createResult.count, 0)
 
         let insertQuery = """
@@ -218,10 +217,10 @@ class PostgreSQLConnectionTests: XCTestCase {
         params += UUID() // new uuid
         // params.append([1,2,3] as [Int]) // new array
 
-        let insertResult = try! client.query(insertQuery, params).await(on: eventLoop)
+        let insertResult = try client.query(insertQuery, params).wait()
         XCTAssertEqual(insertResult.count, 0)
 
-        let parameterizedResult = try! client.query("select * from kitchen_sink").await(on: eventLoop)
+        let parameterizedResult = try client.query("select * from kitchen_sink").wait()
         if parameterizedResult.count == 1 {
             let row = parameterizedResult[0]
             try XCTAssertEqual(row["smallint"]?.decode(Int16.self), 1)
@@ -251,36 +250,36 @@ class PostgreSQLConnectionTests: XCTestCase {
             var message: String
         }
 
-        let (client, eventLoop) = try! PostgreSQLConnection.makeTest()
-        _ = try! client.query("drop table if exists foo;").await(on: eventLoop)
-        let createResult = try! client.query("create table foo (id integer, dict jsonb);").await(on: eventLoop)
+        let client = try PostgreSQLConnection.makeTest()
+        _ = try client.query("drop table if exists foo;").wait()
+        let createResult = try client.query("create table foo (id integer, dict jsonb);").wait()
         XCTAssertEqual(createResult.count, 0)
-        let insertResult = try! client.query("insert into foo values ($1, $2);", [
+        let insertResult = try client.query("insert into foo values ($1, $2);", [
             Int32(1), Hello(message: "hello, world")
-        ]).await(on: eventLoop)
+        ]).wait()
 
         XCTAssertEqual(insertResult.count, 0)
-        let parameterizedResult = try! client.query("select * from foo").await(on: eventLoop)
+        let parameterizedResult = try client.query("select * from foo").wait()
         if parameterizedResult.count == 1 {
             let row = parameterizedResult[0]
-            try! XCTAssertEqual(row["id"]?.decode(Int.self), 1)
-            try! XCTAssertEqual(row["dict"]?.decode(Hello.self).message, "hello, world")
+            try XCTAssertEqual(row["id"]?.decode(Int.self), 1)
+            try XCTAssertEqual(row["dict"]?.decode(Hello.self).message, "hello, world")
         } else {
             XCTFail("parameterized result count is: \(parameterizedResult.count)")
         }
     }
 
     func testNull() throws {
-        let (client, eventLoop) = try PostgreSQLConnection.makeTest()
-        _ = try client.query("drop table if exists nulltest;").await(on: eventLoop)
-        let createResult = try client.query("create table nulltest (i integer not null, d timestamp);").await(on: eventLoop)
+        let client = try PostgreSQLConnection.makeTest()
+        _ = try client.query("drop table if exists nulltest;").wait()
+        let createResult = try client.query("create table nulltest (i integer not null, d timestamp);").wait()
         XCTAssertEqual(createResult.count, 0)
-        let insertResult = try! client.query("insert into nulltest  (i, d) VALUES ($1, $2)", [
+        let insertResult = try client.query("insert into nulltest  (i, d) VALUES ($1, $2)", [
             PostgreSQLData(type: .int2, format: .binary, data: Data([0x00, 0x01])),
             PostgreSQLData(type: .timestamp, format: .binary, data: nil),
-        ]).await(on: eventLoop)
+        ]).wait()
         XCTAssertEqual(insertResult.count, 0)
-        let parameterizedResult = try! client.query("select * from nulltest").await(on: eventLoop)
+        let parameterizedResult = try client.query("select * from nulltest").wait()
         XCTAssertEqual(parameterizedResult.count, 1)
     }
 
@@ -297,13 +296,13 @@ class PostgreSQLConnectionTests: XCTestCase {
 
 extension PostgreSQLConnection {
     /// Creates a test event loop and psql client.
-    static func makeTest() throws -> (PostgreSQLConnection, EventLoop) {
-        let eventLoop = try DefaultEventLoop(label: "codes.vapor.postgresql.client.test")
-        let client = try PostgreSQLConnection.connect(on: eventLoop) { _, error in
+    static func makeTest() throws -> PostgreSQLConnection {
+        let group = MultiThreadedEventLoopGroup(numThreads: 1)
+        let client = try PostgreSQLConnection.connect(on: wrap(group.next())) { error in
             XCTFail("\(error)")
-        }
-        _ = try client.authenticate(username: "postgres").await(on: eventLoop)
-        return (client, eventLoop)
+        }.wait()
+        _ = try client.authenticate(username: "postgres").wait()
+        return client
     }
 }
